+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+*
+
+### Changed
+
+*
+
+### Fixed
+
+*
+
+## [0.16.0] - 2022-02-11
+
+### Added
+
 * Added `source` prop to `Lightbox`
 
 ### Changed